split chat component, added premium logic
