+++ frontend/src/components/layouts/main-layout.tsx
@@ -1,4 +1,4 @@
-import { ChatBubble } from '@/components/chat/ChatBubble';
+import { Chat } from '@/components/chat/chat';
 import { Header } from '@/components/header/header';
 import { TutorialProvider } from '@/contexts/tutorial-context';
 import type { PropsWithChildren } from 'react';
@@ -10,7 +10,7 @@ export function MainLayout({ children }: PropsWithChildren) {
         <Header />
         <main className="mx-auto max-w-7xl">{children}</main>
       </div>
-      <ChatBubble />
+      <Chat />
     </TutorialProvider>
   );
 }
